requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.8.1
numpy>=1.24.0
matplotlib>=3.7.0
pydantic-settings>=2.0.0
e2b>=0.1.0
docker>=6.0.0
python-dotenv>=1.0.0
openai>=1.0.0  # For OpenAI model support
anthropic>=0.18.0  # For Anthropic model support
google-generativeai>=0.3.0  # For Gemini model support
asyncio>=3.4.3  # For async/await support 
jinja2>=3.1.0  # For Jinja2 templating
requests>=2.20.0 # For making HTTP requests